+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.5.0] - 2022-12-13
+### Added
+- The function `DeleteFlightLog` to interact with the flight log service to
+  delete a specific flight log.
+- The `DELETE /flight-log/-` endpoint to handle the deletion of a specific
+  flight log.
+
 ## [0.4.0] - 2022-12-13
 ### Added
 - The function `UpdateFlightLog` to interact with the flight log microservice